requirements.txt
tencentcloud-sdk-python
requests
Flask
Pillow
